+++ .gitignore
@@ -24,3 +24,4 @@ pnpm-debug.log*
 
 .firebase/
 .runtimeconfig.json
+TODO.md
