+++ README.md
@@ -21,5 +21,3 @@ npm run report
 # Compiles and minifies for production
 npm run build
 ```
-
-react version [here](https://github.com/vinyl-umbrella/futon-openchat_web_app/tree/react-rewrite)
