+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "futon-openchat",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve --port 8888",
@@ -10,27 +10,25 @@
   },
   "dependencies": {
     "animejs": "^3.2.1",
-    "chart.js": "^2.9.4",
-    "core-js": "^3.19.1",
-    "hls.js": "^1.1.1",
-    "vue": "^2.6.14",
-    "vue-chartjs": "^3.5.1",
-    "vue-router": "^3.5.3",
-    "vuetify": "^2.4.0"
+    "chart.js": "^3.7.0",
+    "core-js": "^3.6.5",
+    "hls.js": "^1.1.3",
+    "primeicons": "^5.0.0",
+    "primevue": "^3.11.0",
+    "vue": "^3.2.22",
+    "vue-router": "^4.0.0-0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.15",
-    "@vue/cli-plugin-eslint": "~4.5.15",
-    "@vue/cli-plugin-router": "~4.5.15",
-    "@vue/cli-service": "^4.5.15",
+    "@vue/cli-plugin-babel": "~4.5",
+    "@vue/cli-plugin-eslint": "~4.5",
+    "@vue/cli-plugin-router": "~4.5",
+    "@vue/cli-service": "~4.5",
+    "@vue/compiler-sfc": "^3.0.0",
+    "@vue/eslint-config-prettier": "^6.0.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "sass": "~1.32.12",
-    "sass-loader": "^7.1.0",
-    "vue-cli-plugin-vuetify": "~2.4.3",
-    "vue-template-compiler": "^2.6.14",
-    "vuetify-loader": "^1.7.3",
-    "webpack-bundle-analyzer": "4.5"
+    "eslint": "^6.7.2",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-vue": "^7.0.0",
+    "prettier": "^2.2.1"
   }
 }
