--- src/plugins/vuetify.js
@@ -1,22 +1,0 @@
-import Vue from 'vue';
-import Vuetify from 'vuetify/lib';
-import 'vuetify/dist/vuetify.min.css';
-
-Vue.use(Vuetify);
-
-export default new Vuetify({
-  theme: {
-      dark: true,
-      options: {
-        customProperties: true,
-      },
-    themes: {
-      dark: {
-        primary: '#ddd',
-        secondary: '#03A9F4',
-        info: '#44424e',
-        background: '#44424e',
-      }
-    },
-  },
-});
