+++ src/views/Contact.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    アマギフ，苦情，その他管理人に伝えたいことは
+    <span>アマギフ，苦情，その他管理人に伝えたいことは</span>
     <a
       href="https://docs.google.com/forms/d/1-UvCM5TZJ2Syfi1Cm5NKySgF8I65QixlbwC0_mgdrlo/viewform"
       target="_blank"
@@ -11,9 +11,3 @@
     から
   </div>
 </template>
-
-<style scoped>
-a:hover {
-  color: var(--v-secondary-base);
-}
-</style>
