+++ src/views/Stream.vue
@@ -1,4 +1,4 @@
-<template>
+<!-- <template>
   <div>
     <div class="flexbox">
       <v-text-field
@@ -214,4 +214,12 @@ export default {
   position: relative;
   overflow: hidden;
 }
-</style>
+</style> -->
+
+<template>
+  <div>hello!</div>
+</template>
+
+<script setup></script>
+
+<style></style>
