--- vue3/README.md
@@ -1,23 +1,0 @@
-# openrec-tools
-## features
-- masaoroid
-- openrec openchat ranking of number of posts
-- openrec openchat backlog search
-- openrec comment viewer
-- openrec player like niconico
-- openrec.tv api doc
-
-## dev
-### hosting
-```sh
-# install package
-npm install
-# local server
-npm run serve
-# lint
-npm run lint
-# make report
-npm run report
-# Compiles and minifies for production
-npm run build
-```
