--- vue3/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "vue3",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve --port 8888",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "report": "vue-cli-service build --report"
-  },
-  "dependencies": {
-    "chart.js": "^3.7.0",
-    "core-js": "^3.6.5",
-    "hls.js": "^1.1.3",
-    "primeicons": "^5.0.0",
-    "primevue": "^3.11.0",
-    "vue": "^3.2.22",
-    "vue-router": "^4.0.0-0"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5",
-    "@vue/cli-plugin-eslint": "~4.5",
-    "@vue/cli-plugin-router": "~4.5",
-    "@vue/cli-service": "~4.5",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^7.0.0",
-    "prettier": "^2.2.1"
-  }
-}
