--- vue3/src/views/Contact.vue
@@ -1,13 +1,0 @@
-<template>
-  <div>
-    <span>アマギフ，苦情，その他管理人に伝えたいことは</span>
-    <a
-      href="https://docs.google.com/forms/d/1-UvCM5TZJ2Syfi1Cm5NKySgF8I65QixlbwC0_mgdrlo/viewform"
-      target="_blank"
-      rel=" noopener norefferer"
-    >
-      google form
-    </a>
-    から
-  </div>
-</template>
