--- vue3/src/views/Home.vue
@@ -1,99 +1,0 @@
-<template>
-  <div class="masaoroid">
-    <h1>Masaoroid</h1>
-    <div>
-      <div class="masao-msg">
-        <p>{{ message }}</p>
-        <p class="warn">※本人の発言ではありません</p>
-      </div>
-      <div class="masao-face">( ･᷄෴･᷅.)</div>
-    </div>
-    <Button
-      :loading="nowloading"
-      label="再生成"
-      class="p-button-outlined"
-      @click="getMasaoMessage()"
-    />
-  </div>
-</template>
-
-<script setup>
-import { ref, onMounted } from "vue";
-import { useToast } from "primevue/usetoast";
-
-const message = ref("loading...");
-const nowloading = ref(true);
-const toast = useToast();
-
-onMounted(() => {
-  getMasaoMessage();
-});
-
-const getMasaoMessage = async () => {
-  let url =
-    "https://asia-northeast1-futonchan-openchat.cloudfunctions.net/masaoroid";
-  nowloading.value = true;
-  let res = await fetch(url);
-  nowloading.value = false;
-  if (res.ok) {
-    message.value = await res.text();
-  } else {
-    toast.add({
-      severity: "error",
-      summary: "Failed",
-      detail: "Failed to get msg",
-      life: 3000,
-    });
-  }
-};
-</script>
-
-<style scoped>
-.masaoroid {
-  position: fixed;
-  width: 80vw;
-  height: 80vh;
-  top: 50%;
-  left: 50%;
-  margin-left: -40vw;
-  margin-top: -20vh;
-  text-align: center;
-}
-
-.masaoroid > div {
-  font-size: 30px;
-}
-
-.masao-msg {
-  position: relative;
-  display: inline-block;
-  background-color: transparent;
-  border: solid 5px var(--primary-color);
-  padding: 16px;
-  min-width: 240px;
-  width: 50%;
-  text-align: center;
-  border-radius: 5px;
-}
-.masao-msg::before {
-  content: "";
-  position: absolute;
-  border: solid 12px transparent;
-  border-top: solid 12px var(--primary-color);
-  top: 100%;
-  left: 50%;
-  -webkit-transform: translateX(-50%);
-  transform: translateX(-50%);
-}
-.masao-msg p {
-  margin: 0;
-}
-.masao-msg .warn {
-  font-size: 10px;
-  text-align: left;
-}
-
-.masao-face {
-  margin: 10px 0px;
-}
-</style>
