--- vue3/src/views/Stream.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>hello!</div>
-</template>
-
-<script setup></script>
-
-<style></style>
